# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Ecommerce Backend API
  summary: Production-level e-commerce API (storefront + admin console).
  description: 'Session-cookie authenticated REST API. Money values are decimal strings

    (`"10.00"`, pattern `^\d{1,10}(\.\d{1,2})?$`); timestamps are ISO 8601 UTC

    (`date-time` with offset); only public IDs (`usr_`, `prd_`, …) are exposed.

    All write operations require a session cookie and, when CSRF is enabled, an

    `x-csrf-token` header.


    **Authentication:** Opaque DB-backed `session` HttpOnly cookie (SameSite=Lax).

    Obtain via `POST /auth/register` or `POST /auth/login`; revoke via

    `DELETE /auth/session`. Call `GET /auth/csrf-token` before cookie-authenticated

    writes.


    **Pagination:** `page` (1-based, default 1) + `limit` (1–100, default 20).

    Response `meta` carries `total`, `totalPages`, `hasNext`, `hasPrev`.


    **Source of truth:** the hand-maintained contracts in `docs/api/**` remain

    authoritative for semantics; this file is a machine-importable projection of

    them (kept in sync manually — see openapi/README.md). Markdown wins on any

    conflict until this note is reversed.


    **Password policy:** minimum 8 characters and must include an uppercase

    letter, a lowercase letter, a number and a special character

    (`/(?=.*[a-z])(?=.*[A-Z])(?=.*\d)(?=.*[\W_]).{8,}/`).

    '
  version: 1.0.0
  contact:
    name: API Support
    url: https://github.com/Abo3baziz/ecommerce-api
...
servers:
...
security:
- sessionCookie: []
paths:
...
  /admin/settings/email/test:
    post:
      tags:
      - Admin � System Settings
      summary: Send a test email using the configured provider
      description: Acknowledges a test send to the given address. Requires SUPER_ADMIN.
      operationId: adminTestSettingsEmail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - to
              properties:
                to:
                  type: string
                  format: email
      responses:
        '200':
          description: Test email acknowledged.
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/SuccessEnvelope'
                - type: object
                  properties:
                    data:
                      type: object
                      properties:
                        sent_to:
                          type: string
                          format: email
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /admin/admins:
    get:
      tags:
      - Admin — Admins
      summary: List admin accounts
      description: Lists admin accounts with activity aggregates for SUPER_ADMIN, supporting pagination, search, status (ACTIVE, SUSPENDED) and activity (ACTIVE, INACTIVE based on recent login/audit within 2 days) filters, and sorting.
      operationId: adminListAdmins
      parameters:
      - $ref: '#/components/parameters/PageParam'
      - $ref: '#/components/parameters/LimitParam'
      - $ref: '#/components/parameters/SearchParam'
      - name: status
        in: query
        required: false
        schema:
          type: string
          enum:
          - ACTIVE
          - SUSPENDED
      - name: activity
        in: query
        required: false
        schema:
          type: string
          enum:
          - ACTIVE
          - INACTIVE
          description: ACTIVE means recent login/audit activity within 2 days.
      - name: sort
        in: query
        required: false
        schema:
          type: string
          enum:
          - name
          - -name
          - created_at
          - -created_at
          - last_login_at
          - -last_login_at
          default: -last_login_at
      responses:
        '200':
          description: OK.
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/SuccessEnvelope'
                - type: object
                  properties:
                    data:
                      type: array
                      items:
                        type: object
                    meta:
                      $ref: '#/components/schemas/PaginationMeta'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/admins/{admin_public_id}:
    get:
      tags:
      - Admin — Admins
      summary: Get admin details
      description: Returns detailed activity information for a single admin account, including recent logins and audited actions. Requires SUPER_ADMIN role.
      operationId: adminGetAdmin
      parameters:
      - name: admin_public_id
        in: path
        required: true
        schema:
          type: string
          minLength: 1
          example: usr_01H
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
      description: Opaque DB-backed session token (HttpOnly, SameSite=Lax).
  schemas:
...
    SuccessEnvelope:
      type: object
      required:
      - success
      properties:
        success:
          type: boolean
          const: true
          example: true
      description: Base success envelope; extended via allOf with `data` and `meta`.
    ErrorEnvelope:
      type: object
      required:
      - success
      - message
      properties:
        success:
          type: boolean
          const: false
          example: false
        message:
          type: string
          example: Invalid request
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                example: email
              message:
                type: string
                example: Invalid email
        requestId:
          type: string
          example: req_01H
    PaginationMeta:
      type: object
      required:
      - page
      - limit
      - total
      properties:
        page:
          type: integer
          minimum: 1
          example: 1
        limit:
          type: integer
          minimum: 1
          example: 20
        total:
          type: integer
          minimum: 0
          example: 42
        totalPages:
          type: integer
          minimum: 0
          example: 3
        hasNext:
          type: boolean
          example: true
        hasPrev:
          type: boolean
          example: false
...
  parameters:
    PageParam:
      name: page
      in: query
      required: false
      description: Page number (1-based).
      schema:
        type: integer
        minimum: 1
        default: 1
        example: 1
    LimitParam:
      name: limit
      in: query
      required: false
      description: Items per page (1–100).
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
        example: 20
    SearchParam:
      name: search
      in: query
      required: false
      description: Free-text search (trimmed).
      schema:
        type: string
        maxLength: 100
        example: sneaker
...
  responses:
    BadRequest:
      description: Invalid body, query or path parameter.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            success: false
            message: Validation failed
            errors:
            - field: email
              message: Invalid email
            requestId: req_01H
    Unauthorized:
      description: Missing, invalid or expired session.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            success: false
            message: Authentication required
            requestId: req_01H
    Forbidden:
      description: Authenticated but insufficient role, or account suspended.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            success: false
            message: Forbidden
            requestId: req_01H
    NotFound:
      description: Resource does not exist (or is deliberately hidden).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            success: false
            message: Resource not found
            requestId: req_01H
...